Move all headers to include/fastnumbers.
This forces all includes to use #include "fastnumbers/strings.h",
which significantly reduces the chance of name conflicts.

On the FreeBSD port, there was a conflict with fastnumbers' "strings.h"
and the system "strings.h" which caused compilation to fail - this
is the cause of issue #18. It is hoped this commit will solve issue #18.
