+++ include/fastnumbers/fn_bool.h
