+++ src/parsing.c
@@ -6,8 +6,8 @@
  */
 #include <Python.h>
 #include <string.h>
-#include "parsing.h"
-#include "pstdint.h"
+#include "fastnumbers/parsing.h"
+#include "fastnumbers/pstdint.h"
 
 /******************/
 /****** NOTE ******/
